LF-4164 Move origin logic into Origin component
