+++ packages/webapp/src/stories/Animals/Details/AnimalDetails.stories.tsx
@@ -20,7 +20,7 @@ import AnimalCreationDetails, {
   AnimalDetailsProps,
 } from '../../../components/Animals/AddAnimalsDetails';
 import AnimalDetails from '../../../components/Animals/AddAnimalsDetails';
-import { DetailsFields, FormMethods } from '../../../containers/Animals/AddAnimals/types';
+import { FormMethods } from '../../../containers/Animals/AddAnimals/types';
 import {
   sexOptions,
   useOptions,
@@ -30,7 +30,6 @@ import {
   originOptions,
   defaultValues,
 } from './mockData';
-import { AnimalOrigins } from '../../../containers/Animals/types';
 
 // https://storybook.js.org/docs/writing-stories/typescript
 const meta: Meta<AnimalDetailsProps> = {
@@ -46,13 +45,6 @@ export const Default: Story = {
     const formMethods: FormMethods = useForm({
       defaultValues,
     });
-
-    const originId = formMethods.watch(DetailsFields.ORIGIN);
-    const origin = !originId
-      ? undefined
-      : originId === 1
-        ? AnimalOrigins.BROUGHT_IN
-        : AnimalOrigins.BORN_AT_FARM;
 
     return (
       <Suspense>
@@ -73,7 +65,6 @@ export const Default: Story = {
               originProps={{
                 currency: '$',
                 originOptions,
-                origin,
               }}
             />
           </FormProvider>
